split main.py into main.py + addon.py
